Split out factories
Split out existing code to create the `IRuleFactory`, `IPropertyPagesCatalogFactory`, and `IPropertyPagesCatalogProviderFactory` as I'm going to need to create mocks of the related types for the property page work.
