+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/IPropertyPagesCatalogFactory.cs
@@ -1,0 +1,67 @@
+﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.VisualStudio.ProjectSystem.Properties;
+using Moq;
+
+namespace Microsoft.VisualStudio.ProjectSystem
+{
+    internal static class IPropertyPagesCatalogFactory
+    {
+        public static IPropertyPagesCatalog Create(Dictionary<string, IRule> rulesBySchemaName)
+        {
+            var catalog = new Mock<IPropertyPagesCatalog>();
+            catalog.Setup(o => o.BindToContext(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
+                   .Returns((string schemaName, string file, string itemType, string itemName) =>
+                   {
+
+                       rulesBySchemaName.TryGetValue(schemaName, out IRule rule);
+                       return rule;
+                   });
+
+            return catalog.Object;
+        }
+
+        public static IPropertyPagesCatalog Create(params PropertyPageData[] data)
+        {
+            return Create(CreateCatalogLookup(data));
+        }
+
+        private static Dictionary<string, IRule> CreateCatalogLookup(PropertyPageData[] data)
+        {
+            var catalog = new Dictionary<string, IRule>();
+
+            foreach (var category in data.GroupBy(p => p.Category))
+            {
+                catalog.Add(category.Key,
+                            IRuleFactory.Create(
+                                category.Select(property => CreateProperty(property.PropertyName, property.Value, property.SetValues))));
+            }
+
+            return catalog;
+        }
+
+        private static IProperty CreateProperty(string name, object value, List<object>? setValues = null)
+        {
+            var property = new Mock<IProperty>();
+            property.SetupGet(o => o.Name)
+                    .Returns(name);
+
+            property.Setup(o => o.GetValueAsync())
+                    .ReturnsAsync(value);
+
+            property.As<IEvaluatedProperty>().Setup(p => p.GetEvaluatedValueAtEndAsync()).ReturnsAsync(value.ToString());
+            property.As<IEvaluatedProperty>().Setup(p => p.GetEvaluatedValueAsync()).ReturnsAsync(value.ToString());
+
+            if (setValues != null)
+            {
+                property.Setup(p => p.SetValueAsync(It.IsAny<object>()))
+                        .Callback<object>(setValues.Add)
+                        .ReturnsAsync(() => { });
+            }
+
+            return property.Object;
+        }
+    }
+}
