+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/IPropertyPagesCatalogProviderFactory.cs
@@ -1,0 +1,27 @@
+﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Threading;
+using Microsoft.VisualStudio.ProjectSystem.Properties;
+using Moq;
+
+namespace Microsoft.VisualStudio.ProjectSystem
+{
+    internal static class IPropertyPagesCatalogProviderFactory
+    {
+        public static IPropertyPagesCatalogProvider Create(Dictionary<string, IPropertyPagesCatalog> catalogsByContext, IPropertyPagesCatalog memoryOnlyCatalog)
+        {
+            var catalogProvider = new Mock<IPropertyPagesCatalogProvider>();
+            catalogProvider
+                .Setup(o => o.GetCatalogsAsync(CancellationToken.None))
+                .ReturnsAsync(catalogsByContext.ToImmutableDictionary());
+
+            catalogProvider
+                .Setup(o => o.GetMemoryOnlyCatalog(It.IsAny<string>()))
+                .Returns(memoryOnlyCatalog);
+
+            return catalogProvider.Object;
+        }
+    }
+}
