+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/ProjectPropertiesFactory.cs
@@ -1,9 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
-using System.Threading;
 using Microsoft.VisualStudio.ProjectSystem.Properties;
 using Moq;
 
@@ -30,8 +27,8 @@ namespace Microsoft.VisualStudio.ProjectSystem
 
         public static ProjectProperties Create(UnconfiguredProject project, params PropertyPageData[] data)
         {
-            var catalog = CreateCatalog(CreateCatalogLookup(data));
-            IPropertyPagesCatalogProvider propertyPagesCatalogProvider = CreateCatalogProvider(
+            var catalog = IPropertyPagesCatalogFactory.Create(data);
+            var propertyPagesCatalogProvider = IPropertyPagesCatalogProviderFactory.Create(
                     new Dictionary<string, IPropertyPagesCatalog>
                     {
                         { "Project", catalog }
@@ -53,83 +50,5 @@ namespace Microsoft.VisualStudio.ProjectSystem
 
             return new ProjectProperties(configuredProject);
         }
-
-        private static Dictionary<string, IRule> CreateCatalogLookup(PropertyPageData[] data)
-        {
-            var catalog = new Dictionary<string, IRule>();
-
-            foreach (var category in data.GroupBy(p => p.Category))
-            {
-                catalog.Add(category.Key,
-                            CreateRule(
-                                    category.Select(property => CreateProperty(property.PropertyName, property.Value, property.SetValues))));
-            }
-
-            return catalog;
-        }
-
-        private static IPropertyPagesCatalogProvider CreateCatalogProvider(Dictionary<string, IPropertyPagesCatalog> catalogsByContext, IPropertyPagesCatalog catalog)
-        {
-            var catalogProvider = new Mock<IPropertyPagesCatalogProvider>();
-            catalogProvider
-                .Setup(o => o.GetCatalogsAsync(CancellationToken.None))
-                .ReturnsAsync(catalogsByContext.ToImmutableDictionary());
-
-            catalogProvider
-                .Setup(o => o.GetMemoryOnlyCatalog(It.IsAny<string>()))
-                .Returns(catalog);
-
-            return catalogProvider.Object;
-        }
-
-        private static IPropertyPagesCatalog CreateCatalog(Dictionary<string, IRule> rulesBySchemaName)
-        {
-            var catalog = new Mock<IPropertyPagesCatalog>();
-            catalog.Setup(o => o.BindToContext(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
-                   .Returns((string schemaName, string file, string itemType, string itemName) =>
-                   {
-
-                       rulesBySchemaName.TryGetValue(schemaName, out IRule rule);
-                       return rule;
-                   });
-
-            return catalog.Object;
-        }
-
-        private static IRule CreateRule(IEnumerable<IProperty> properties)
-        {
-            var rule = new Mock<IRule>();
-            rule.Setup(o => o.GetProperty(It.IsAny<string>()))
-                .Returns((string propertyName) =>
-                {
-
-                    return properties.FirstOrDefault(p => p.Name == propertyName);
-                });
-
-            return rule.Object;
-        }
-
-        private static IProperty CreateProperty(string name, object value, List<object>? setValues = null)
-        {
-            var property = new Mock<IProperty>();
-            property.SetupGet(o => o.Name)
-                    .Returns(name);
-
-            property.Setup(o => o.GetValueAsync())
-                    .ReturnsAsync(value);
-
-            property.As<IEvaluatedProperty>().Setup(p => p.GetEvaluatedValueAtEndAsync()).ReturnsAsync(value.ToString());
-            property.As<IEvaluatedProperty>().Setup(p => p.GetEvaluatedValueAsync()).ReturnsAsync(value.ToString());
-
-            if (setValues != null)
-            {
-                property.Setup(p => p.SetValueAsync(It.IsAny<object>()))
-                        .Callback<object>(setValues.Add)
-                        .ReturnsAsync(() => { });
-            }
-
-            return property.Object;
-        }
-
     }
 }
